Support link and payloadlink
support link in list response
